Project Started and Completed
Since all the Northwind-Main project is, is to be a place-holder for a public location for documentation, I decided that it should also be where we keep the main application, consisting of only the platform and the modules that are needed to run. Therefore, I created this project and did all of the branding on it, so that Northwind would be ready to go as we get plugins created for it.

-SC