Initial Commit of DBSystem Module
This commit contains only the following Module, packages, and class/files:

- DBSystem Module
   - `com.pekinsoft.nwind.db` - The main database system package.
      - `Bundle.properties` - The I18N properties for the module.
   - `com.pekinsoft.nwind.db.scripts` - The package that provides for connection to the database, as well as database creation, if needed.
      - `ConnectToDatabase.java` - The class that provides the connection and creation of the database.

I am going to create some views into the database, such as a way of viewing all of the tables the database contains, as well as opening each table in a view and being able to read, edit, add, and delete data from the tables. These windows will be located under 'Window/Administration' submenu. These will not be the primary windows for any of the functionality of the Northwind Traders system, just for database administration, such as adding new accounts, or deleting accounts, etc. I am not sure how we will handle this, currently, but am willing to allow it to be done for customization purposes.

-SC