Database System Revision 2
OK, so I have gone a different route here. The class `com.pekinsoft.nwind.db.scripts.ConnectToDatabase` is still present, but may not actually be needed. I will need your input on this @jkovalski.

Now, there is the `META-INF` package at the root level of the source packages, containing the `persistence.xml` file for connecting to the database and generating the tables if needed.

Also, there is now the package `com.pekinsoft.nwind.db.api`, which contains the following classes:

- `AccountTypes` - The types of accounts being used in this accounting system:
   - Expense
   - Revenue
   - Assets
   - Liabilities
- `AccountingJournal` - The journal for accounting expenses
- `Accounts` - The chart of accounts
- `CheckingJournal` - Checking account journal
- `CommunicationJournal` - communications expense journal
- `CreditCardJournal` - credit card liability journal
- `FeesJournal` - Fees expense journal
- `FuelCardJournal` - Fuel Card asset journal
- `FuelJournal` - Fuel expense journal
- `Gl` - The General Ledger for the accounting system
- `InsuranceJournal` - Insurance expense journal
- `InterestJournal` - Interest expense journal
- `LoadingJournal` - Loading/Unloading expense journal
- `MaintenanceJournal` - Maintenance/Repair expense journal
- `MaintenanceTypes` - Types of maintenance that may occur
- `OfficeJournal` - Office supply expense journal
- `PhysicalsJournal` - Business-related medical expense journal
- `RentJournal` - Rental/Lease expense journal
- `ReservesJournal` - Maintenance Reserves asset journal
- `SavingsJournal` - Savings account asset journal
- `ScalesJournal` - Scales expense journal
- `SuppliesJournal` - Supplies and equipment expense journal
- `TaxJournal` - Tax-related expense journal
- `TollsJournal` - Tolls expense journal
- `TravelJournal` - Business-related travel expense journal
- `UniformsJournal` - Business-related uniforms/clothing expense journal
- `WagesJournal` - Payroll expense journal, not to include personal wages

I still need to add the load-related tables, which I just now realized that I had forgotten to do. I will get that added and pushed up on the next Revision of this branch.

-SC